+++ Src/Runtime/Module/Entity/Data/MonsterDataCore.cs
@@ -11,7 +11,7 @@ using UnityGameFramework.Runtime;
 /// </summary>
 public class MonsterDataCore : EntityBaseComponent
 {
-    public int Lv => RefEntity.BattleDataCore.Level;
+    public float Lv => RefEntity.BattleDataCore.Level;
     public int configId => DRMonster == null ? -1 : DRMonster.Id;
     public DRMonster DRMonster { get; protected set; }
